Update PowerShell Profile
Optimised PowerShell Profile script so instead of using Get-Module to check if our favourite modules are already installed (this cmdlet seems to be incredibly slow), we instead create a registry key for each module to track if it was successfully installed at some stage.  This significantly improves PowerShell startup times.